+++ webpack.main.config.js
@@ -10,7 +10,9 @@ module.exports = {
     new CopyWebpackPlugin({
       patterns: [
         { from: 'source/main/assets/tutorial', to: 'tutorial' },
-        { from: 'source/main/assets/dict', to: 'dict' }
+        { from: 'source/main/assets/dict', to: 'dict' },
+        { from: 'source/app/service-providers/assets/csl-locales', to: 'assets/csl-locales' },
+        { from: 'source/app/service-providers/assets/csl-styles', to: 'assets/csl-styles' }
       ]
     })
   ],